+++ codegen/Crm/Lists/Model/PublicRangedNumberPropertyOperation.php
@@ -73,10 +73,10 @@ class PublicRangedNumberPropertyOperation implements ModelInterface, ArrayAccess
       */
     protected static $openAPIFormats = [
         'include_objects_with_no_value_set' => null,
-        'lower_bound' => 'int32',
+        'lower_bound' => 'int64',
         'operation_type' => null,
         'operator' => null,
-        'upper_bound' => 'int32'
+        'upper_bound' => 'int64'
     ];
 
     /**
